Move integrand_schatten_norm from slq.py to lanczos.py
